Add Biquad Filter tests (#2)
* work

* more testing

* fix

* improve github action

* try upload plots

* try again

* ahhh

* try again

* try again

* what is happening

* simplification?

* works i think

* AHHhh

+++ siderialib/src/effects/filter/BiquadFilter.cpp
@@ -52,10 +52,10 @@ void BiquadFilter::setParams(sfloat cutoff, sfloat Q, sfloat dB) {
 }
 
 sfloat BiquadFilter::tick(sfloat x0) {
-    _out = (x0 * _a0) + (_x1 * _a1) +  (_x2 * _a2) - (_y1 * _b1) - (_y2 * _b2);
+    _out = ((double)x0 * _a0) + (_x1 * _a1) +  (_x2 * _a2) - (_y1 * _b1) - (_y2 * _b2);
 
     _x2 = _x1;
-    _x1 = x0;
+    _x1 = (double)x0;
 
     _y2 = _y1;
     _y1 = _out;
